% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaccd.R
\name{getRefCor}
\alias{getRefCor}
\title{Retrieve the reference correlation matrix for clock gene co-expression.}
\usage{
getRefCor(species = "human", useEntrezGeneId = TRUE)
}
\arguments{
\item{species}{Currently either 'human' or 'mouse'. Only affects the row and column names
of the correlation matrix, not the actual values.}

\item{useEntrezGeneId}{If FALSE, row and column names of correlation matrix
will correspond to gene symbols (e.g., PER2).}
}
\value{
A matrix of Spearman correlation values.
}
\description{
The reference matrix is based on a fixed-effects meta-analysis of eight circadian
transcriptome datasets from mice, as described in \url{https://dx.doi.org/10.1101/130765}.
}
\examples{
\dontrun{
library('deltaccd')
library('doParallel')
library('doRNG')

registerDoParallel(cores=2)
set.seed(35813)

refCor = getRefCor()
ccdResult = calcCCD(refCor, GSE19188$emat, GSE19188$groupVec, dopar=TRUE)
deltaCcdResult = calcDeltaCCD(refCor, GSE19188$emat, GSE19188$groupVec, 'non-tumor', dopar=TRUE)

pRef = plotRefHeatmap(refCor)
pTest = plotHeatmap(rownames(refCor), GSE19188$emat, GSE19188$groupVec)
}

}
\seealso{
\code{\link{GSE19188}}, \code{\link{plotRefHeatmap}}, \code{\link{calcCCD}}, \code{\link{calcDeltaCCD}}
}
