% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaccd.R
\docType{data}
\name{GSE19188}
\alias{GSE19188}
\title{Gene expression data for GSE19188.}
\format{
A list with two objects:
\describe{
\item{emat}{Matrix of normalized expression values, where each row
corresponds to a gene (rownames are Entrez Gene IDs) and each column
corresponds to a sample. To save space, genes have been downsampled.}
\item{groupVec}{Vector of condition (tumor or healthy) for each sample.}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE19188}
}
\usage{
GSE19188
}
\description{
Data of gene expression measured by microarray for samples from human
non-small cell lung cancer.
}
\seealso{
\code{\link[=getRefCor]{getRefCor()}}, \code{\link[=calcCCD]{calcCCD()}}, \code{\link[=calcDeltaCCD]{calcDeltaCCD()}}
}
\keyword{datasets}
